# api/nrcnewsapi/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "williamhall2894@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/ciCciC/nrcnewsapi"
        }
    },
    "paths": {
...
        "/category/astronomy/article": {
            "post": {
                "description": "Get a article with article item payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves article",
                "parameters": [
                    {
                        "description": "Get article",
                        "name": "articleitem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArticleItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                }
            }
        },
...
        "/category/games/article": {
            "post": {
                "description": "Get a article with article item payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves article",
                "parameters": [
                    {
                        "description": "Get article",
                        "name": "articleitem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArticleItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                }
            }
        },
...
        "/category/physics/article": {
            "post": {
                "description": "Get a article with article item payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves article",
                "parameters": [
                    {
                        "description": "Get article",
                        "name": "articleitem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArticleItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Article": {
            "type": "object",
            "properties": {
                "imageLink": {
                    "type": "string"
                },
                "pageLink": {
                    "type": "string"
                },
                "sectionList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Section"
                    }
                },
                "teaser": {
                    "type": "string",
                    "example": "BioTelemetry maakt systemen om hartpatiënten te diagnosticeren en op afstand in de gaten te houden. Philips zag de vraag naar zorg op afstand toenemen vanwege de pandemie."
                },
                "title": {
                    "type": "string",
                    "example": "Philips koopt Amerikaanse BioTelemetry voor 2,3 miljard euro"
                },
                "topic": {
                    "type": "string",
                    "example": "E-health"
                }
            }
        },
        "model.ArticleItem": {
            "type": "object",
            "properties": {
                "imageLink": {
                    "type": "string"
                },
                "pageLink": {
                    "type": "string"
                },
                "teaser": {
                    "type": "string",
                    "example": "BioTelemetry maakt systemen om hartpatiënten te diagnosticeren en op afstand in de gaten te houden. Philips zag de vraag naar zorg op afstand toenemen vanwege de pandemie."
                },
                "title": {
                    "type": "string",
                    "example": "Philips koopt Amerikaanse BioTelemetry voor 2,3 miljard euro"
                },
                "topic": {
                    "type": "string",
                    "example": "E-health"
                }
            }
        },
...
        "model.ContentBody": {
            "type": "object",
            "properties": {
                "cType": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "model.Section": {
            "type": "object",
            "properties": {
                "contentBody": {
                    "description": "Contents []string",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContentBody"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}